moving out common logic to utils
